feat: cash in application model fix: type BonusEarned in Level int to float64 feat: add sort by lvl in LevelList feat: calculating invited users & bonus earned by levels for user feat: add BeginAt & ExpireAt in Promo feat: GetBonusTransactionsForUser in TransactionStore feat: add new columns in Supplier (auto upgrade database) feat: new params Status

+++ src/im/model/level.go
@@ -17,8 +17,8 @@ type Level struct {
 	DeleteAt int64   `json:"delete_at"`
 	AppId    string  `json:"app_id"`
 
-	Invited     int `db:"-" json:"invited"`
-	BonusEarned int `db:"-" json:"bonus_earned"`
+	Invited     int     `db:"-" json:"invited"`
+	BonusEarned float64 `db:"-" json:"bonus_earned"`
 }
 
 type LevelPatch struct {
